Change IGV Browser operation behaviour
This commit changes the IGV Browser operation behaviour so that it is opened as a
completely separate application from DEWE. To do this, IGV's log is not shown in
DEWE's log and the progress dialog of the corresponding AIBench operation is not
shown while IGV is opened. Also, the IGV operation is disabled while IGV is opened.

This commit also renames IGV related classes by changing "IGV" with "Igv" to follow
the naming convention.